backend/requirement.txt
aiohttp==3.9.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
async-timeout==4.0.3
attrs==25.4.0
backports-datetime-fromisoformat==2.0.3
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colored==2.2.5
colorful==0.5.8
cryptography==42.0.8
environs==10.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.11
marshmallow==4.2.0
merge-args==0.1.5
multidict==6.7.0
prettyprinter==0.18.0
propcache==0.4.1
psycopg2==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
python-dotenv==1.2.1
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicore==0.2.3
uvicorn==0.40.0
yarl==1.22.0
